# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fisher Fans API",
    "description": "API for managing users, boats, trips, reservations, and logs.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.fisherfans.local/api",
      "description": "Secure local development server"
    }
  ],
  "paths": {
    "/v1/users": {
      "get": {
        "summary": "Get the list of users",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lastName",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "firstName",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "individual",
                "professional"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Edit a user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User edited"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "tags": [
          "Users"
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v1/trips": {
      "get": {
        "summary": "Get the list of trips",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "Filter trips by user",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tripType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of trips",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new trip",
        "tags": [
          "Trips"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trip"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trip created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/trips/{tripId}": {
      "parameters": [
        {
          "name": "tripId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Get a trip by ID",
        "tags": [
          "Trips"
        ],
        "responses": {
          "200": {
            "description": "Trip details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Edit a trip",
        "tags": [
          "Trips"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trip"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trip edited"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a trip",
        "tags": [
          "Trips"
        ],
        "responses": {
          "204": {
            "description": "Trip deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/reservations": {
      "get": {
        "summary": "Get the list of reservations",
        "tags": [
          "Reservations"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "Filter reservations by user",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tripId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reservations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Reservation"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new reservation",
        "tags": [
          "Reservations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reservation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/reservations/{reservationId}": {
      "parameters": [
        {
          "name": "reservationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Get a reservation by ID",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "200": {
            "description": "Reservation details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Edit a reservation",
        "tags": [
          "Reservations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reservation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservation edited"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete a reservation",
        "tags": [
          "Reservations"
        ],
        "responses": {
          "204": {
            "description": "Reservation deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Unprocessable entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the user"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "boatLicense": {
            "type": "string",
            "description": "Boat license number (8 digits)"
          },
          "status": {
            "type": "string",
            "description": "Status of the user (individual or professional)",
            "enum": [
              "individual",
              "professional"
            ]
          },
          "companyName": {
            "type": "string",
            "description": "Company name (empty if individual)"
          },
          "activityType": {
            "type": "string",
            "description": "Type of activity (rental or fishing guide)",
            "enum": [
              "rental",
              "fishing guide"
            ]
          },
          "siretNumber": {
            "type": "string",
            "description": "SIRET number"
          },
          "rcNumber": {
            "type": "string",
            "description": "Commercial register number (RC)"
          },
          "boats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Boat"
            }
          },
          "trips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Trip"
            }
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            }
          },
          "log": {
            "$ref": "#/components/schemas/Log"
          }
        }
      },
      "Boat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the boat"
          },
          "name": {
            "type": "string",
            "description": "Name of the boat"
          },
          "description": {
            "type": "string",
            "description": "Description of the boat"
          },
          "brand": {
            "type": "string",
            "description": "Brand of the boat"
          },
          "manufactureYear": {
            "type": "string",
            "format": "date",
            "description": "Year the boat was manufactured"
          },
          "photoUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL of the boat's photo"
          },
          "licenseType": {
            "type": "string",
            "description": "Required license type (coastal or river)",
            "enum": [
              "coastal",
              "river"
            ]
          },
          "boatType": {
            "type": "string",
            "description": "Type of boat",
            "enum": [
              "open",
              "cabin",
              "catamaran",
              "sailboat",
              "jet ski",
              "canoe"
            ]
          },
          "equipment": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "sonar",
                "livewell",
                "ladder",
                "GPS",
                "rod holders",
                "VHF radio"
              ]
            },
            "description": "List of equipment on board (e.g., sonar, livewell, ladder, GPS, rod holders, VHF radio)"
          },
          "depositAmount": {
            "type": "number",
            "format": "float",
            "description": "Deposit amount (in euros)"
          },
          "maxCapacity": {
            "type": "integer",
            "description": "Maximum number of passengers"
          },
          "numberOfBeds": {
            "type": "integer",
            "description": "Number of beds on the boat"
          },
          "homePort": {
            "type": "string",
            "description": "Home port (city name)"
          },
          "latitude": {
            "type": "number",
            "format": "float",
            "description": "Latitude of the home port"
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "description": "Longitude of the home port"
          },
          "engineType": {
            "type": "string",
            "description": "Type of engine (diesel, gasoline, or none)",
            "enum": [
              "diesel",
              "gasoline",
              "none"
            ]
          },
          "enginePower": {
            "type": "integer",
            "description": "Engine power (in horsepower)"
          }
        }
      },
      "Trip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the trip"
          },
          "title": {
            "type": "string",
            "description": "Title of the trip"
          },
          "practicalInfo": {
            "type": "string",
            "description": "Practical information about the trip"
          },
          "tripType": {
            "type": "string",
            "description": "Type of trip (daily or recurring)",
            "enum": [
              "daily",
              "recurring"
            ]
          },
          "rateType": {
            "type": "string",
            "description": "Type of rate (total or per person)",
            "enum": [
              "total",
              "per person"
            ]
          },
          "startDates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "description": "List of start dates for the trip"
          },
          "endDates": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "description": "List of end dates for the trip"
          },
          "departureTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "time"
            },
            "description": "List of departure times"
          },
          "endTimes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "time"
            },
            "description": "List of end times"
          },
          "passengerCount": {
            "type": "integer",
            "description": "Number of passengers"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Price of the trip"
          },
          "user": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              },
              {
                "description": "User organizing the trip"
              }
            ]
          },
          "boat": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Boat"
              },
              {
                "description": "Boat used for the trip"
              }
            ]
          }
        }
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the reservation"
          },
          "trip": {
            "$ref": "#/components/schemas/Trip"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "reservedSeats": {
            "type": "integer"
          },
          "totalPrice": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "Log": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the log"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Page"
            }
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the page entry"
          },
          "fishName": {
            "type": "string",
            "description": "Name of the fish"
          },
          "fishPhotoUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL of the fish photo"
          },
          "comment": {
            "type": "string",
            "description": "Comment about the catch"
          },
          "length": {
            "type": "number",
            "description": "Length of the fish in centimeters"
          },
          "weight": {
            "type": "number",
            "description": "Weight of the fish in kilograms"
          },
          "fishingSpot": {
            "type": "string",
            "description": "Fishing location"
          },
          "fishingDate": {
            "type": "string",
            "format": "date",
            "description": "Date of fishing"
          },
          "release": {
            "type": "boolean",
            "description": "Was the fish released? (true = yes, false = no)"
          },
          "owner": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              },
              {
                "description": "Owner of the fishing logbook"
              }
            ]
          }
        }
      }
    }
  },
...
}